+++ Backend/tests/activities.test.ts
@@ -1,10 +1,9 @@
-import { api, registerUser, userToRegister} from "../generic_helpers";
+import { api, setUp} from "./generic_helpers";
 import mongoose from "mongoose";
-import Calendar from "../../src/models/calendar.model";
-import Invitation from "../../src/models/invitation.model";
-import User from "../../src/models/user.model";
-import Todo from "../../src/models/tasktoDo.model";
-import { URI, setUp } from "./utils";
+import Calendar from "../src/models/calendar.model";
+import Invitation from "../src/models/invitation.model";
+import User from "../src/models/user.model";
+import Todo from "../src/models/tasktoDo.model";
 
 afterAll(() => {
   mongoose.disconnect();
@@ -16,6 +15,8 @@ beforeEach(async () => {
   await Invitation.deleteMany({});
   await Todo.deleteMany({})
 });
+
+const URI = "/api/activities/"
 
 describe("/api/activities/", () => {
   describe("GET /", () => {
@@ -35,8 +36,8 @@ describe("/api/activities/", () => {
       })
 
       test("Should return an empty list if the user have no activities", async()=>{
-        const userToken = await registerUser(userToRegister[2])
-        const resp = await api.get(URI).set("Authorization", userToken)
+        const {userNoMemberTk} = await setUp()
+        const resp = await api.get(URI).set("Authorization", userNoMemberTk)
 
         expect(resp.body).toBeInstanceOf(Array)
         expect(resp.body).toHaveLength(0)
@@ -50,7 +51,7 @@ describe("/api/activities/", () => {
         const {tokenUser, createdActivity} = await setUp()
         const mondayFstToDo = createdActivity.mondays[0]
 
-        const resp = await api.get(`${URI}/activity/${createdActivity._id}/${mondayFstToDo}`).set("Authorization", tokenUser)
+        const resp = await api.get(`${URI}/${createdActivity._id}/${mondayFstToDo}`).set("Authorization", tokenUser)
 
         expect(resp.statusCode).toBe(200)
       })
@@ -59,7 +60,7 @@ describe("/api/activities/", () => {
         const {tokenUser, createdActivity} = await setUp()
         const mondayFstToDo = createdActivity.mondays[0]
 
-        const resp = await api.get(`${URI}/activity/${createdActivity._id}/${mondayFstToDo}`).set("Authorization", tokenUser)
+        const resp = await api.get(`${URI}/${createdActivity._id}/${mondayFstToDo}`).set("Authorization", tokenUser)
 
         expect(resp.body).toBeInstanceOf(Object)
       })
@@ -67,11 +68,10 @@ describe("/api/activities/", () => {
 
     describe("When other user try to see the Todo", () => {
       test("Should return 400", async()=>{
-        const {tokenUser, createdActivity} = await setUp()
-        const user2 = await registerUser(userToRegister[2])
+        const {userNoMemberTk,  createdActivity} = await setUp()
         const mondayFstToDo = createdActivity.mondays[0]
 
-        const resp = await api.get(`${URI}/activity/${createdActivity._id}/${mondayFstToDo}`).set("Authorization", user2)
+        const resp = await api.get(`${URI}/${createdActivity._id}/${mondayFstToDo}`).set("Authorization", userNoMemberTk)
 
         expect(resp.statusCode).toBe(400)
       })
@@ -84,7 +84,7 @@ describe("/api/activities/", () => {
         const {tokenUser, createdActivity} = await setUp()
         const mondayFstToDo = createdActivity.mondays[0]
 
-        const resp = await api.post(`${URI}/activity/${createdActivity._id}/${mondayFstToDo}/done`).set("Authorization", tokenUser)
+        const resp = await api.post(`${URI}/${createdActivity._id}/${mondayFstToDo}/done`).set("Authorization", tokenUser)
 
         expect(resp.statusCode).toBe(200)
       })
@@ -96,7 +96,7 @@ describe("/api/activities/", () => {
         const todoBefore = await Todo.findById(mondayFstToDo)
         expect(todoBefore!.done).toBeFalsy()
 
-        await api.post(`${URI}/activity/${createdActivity._id}/${mondayFstToDo}/done`).set("Authorization", tokenUser)
+        await api.post(`${URI}/${createdActivity._id}/${mondayFstToDo}/done`).set("Authorization", tokenUser)
 
         const todoAfter = await Todo.findById(mondayFstToDo)
         expect(todoAfter!.done).toBeTruthy()
@@ -107,13 +107,13 @@ describe("/api/activities/", () => {
         const mondayFstToDo = createdActivity.mondays[0]
 
         //First Done
-        await api.post(`${URI}/activity/${createdActivity._id}/${mondayFstToDo}/done`).set("Authorization", tokenUser)
+        await api.post(`${URI}/${createdActivity._id}/${mondayFstToDo}/done`).set("Authorization", tokenUser)
 
         const todoBefore = await Todo.findById(mondayFstToDo)
         expect(todoBefore!.done).toBeTruthy()
       
         //Done again
-        await api.post(`${URI}/activity/${createdActivity._id}/${mondayFstToDo}/done`).set("Authorization", tokenUser)
+        await api.post(`${URI}/${createdActivity._id}/${mondayFstToDo}/done`).set("Authorization", tokenUser)
 
         const todoAfter = await Todo.findById(mondayFstToDo)
         expect(todoAfter!.done).toBeFalsy()
@@ -123,11 +123,10 @@ describe("/api/activities/", () => {
 
     describe("When other user try to see the Todo", () => {
       test("Should return 400", async()=>{
-        const {tokenUser, createdActivity} = await setUp()
-        const user2 = await registerUser(userToRegister[2])
+        const {userNoMemberTk, createdActivity} = await setUp()
         const mondayFstToDo = createdActivity.mondays[0]
 
-        const resp = await api.post(`${URI}/activity/${createdActivity._id}/${mondayFstToDo}/done`).set("Authorization", user2)
+        const resp = await api.post(`${URI}/${createdActivity._id}/${mondayFstToDo}/done`).set("Authorization", userNoMemberTk)
 
         expect(resp.statusCode).toBe(400)
       })
