--- Backend/tests/auth_Tests/utils.ts
@@ -1,26 +1,0 @@
-export const register_initialState = [
-  {
-    name: "Pepe",
-    email: "Pepe@gmail.com",
-    password: "123123",
-    repited_password: "123123",
-  },
-  {
-    name: "Ivan",
-    email: "Ivan@gmail.com",
-    password: "123123",
-    repited_password: "123123",
-  },
-  {
-    name: "Sofia",
-    email: "Sofia@gmail.com",
-    password: "123123",
-    repited_password: "123123",
-  },
-  {
-    name: "Sandra",
-    email: "Sandra@gmail.com",
-    password: "123123",
-    repited_password: "123123",
-  },
-];
