--- Backend/tests/calendar_Tests/utils.ts
@@ -1,18 +1,0 @@
-export const URI = "/api/calendar";
-
-export const cases = [
-  {
-    title: "Title1",
-  },
-  {
-    title: "Title2",
-    description: "Description",
-  },
-  {
-    title: "Title3",
-    description: "Description2",
-  },
-];
-
-export const badCases = [{}, { description: "asdasdasd" }];
-
