--- Backend/tests/invitation_Tests/utils.ts
@@ -1,3 +1,0 @@
-export const URI = "/api/invitations/";
-export const URIcalendar = "/api/calendar/";
-
