--- Backend/tests/user_Tests/utils.ts
@@ -1,1 +1,0 @@
-export const URI = "/api/user";
